% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_util.R
\name{add_luminance}
\alias{add_luminance}
\title{Adding luminance information
Extract and append luminance paths to the data.}
\usage{
add_luminance(obj, files, imgdir = NULL, resize_img = NULL,
  smooth_img = FALSE, verbose = FALSE)
}
\arguments{
\item{obj}{[edar_data] obj, created by load_data or read_ascii.}

\item{files}{[list] of file-paths ordered by trial and 
with names corresponding to the subject IDs in 'obj'.}
}
\description{
Adding luminance information
Extract and append luminance paths to the data.
}

