% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_preprocessing.R
\name{check_preprocessing}
\alias{check_preprocessing}
\title{Function to validate the blink and gaze interpolation
and other stuff to come.}
\usage{
check_preprocessing(obj, what = c("gaze", "ps"), save_pdf = "")
}
\description{
Function to validate the blink and gaze interpolation
and other stuff to come.
}

