% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/commonSingVar.R
\name{commonSingVar}
\alias{commonSingVar}
\title{commonSingVar}
\usage{
commonSingVar(x, trim = FALSE, max_list = getOption("max.print"),
  var_name = "def_x")
}
\arguments{
\item{x}{}

\item{trim}{}

\item{max_list}{}

\item{var_name}{}
}
\value{
output
}
\description{
commonSingVar
}
\examples{
i dunno
}

