% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{describe_cat_var}
\alias{describe_cat_var}
\title{Finds the frequency of the categorical variables in a data frame and returns the histogram of each categorical variable.}
\usage{
describe_cat_var(dataframe, cat_vars)
}
\arguments{
\item{dataframe}{tbl. The dataframe to be inspected.}

\item{cat_vars}{vector of character strings of the names of the categorical variables.}
}
\value{
ggplot object to plot histogram of the categorical variables
}
\description{
Finds the frequency of the categorical variables in a data frame and returns the histogram of each categorical variable.
}
\examples{
X <- dplyr::tibble(type = c('Car','Bus','Car'), height = c(10,20,30))
cat_vars <- c('type')
describe_cat_var(X, cat_vars)

}
