% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{describe_na_values}
\alias{describe_na_values}
\title{Evaluates a dataframe for NA values.}
\usage{
describe_na_values(dataframe)
}
\arguments{
\item{dataframe}{the dataframe to be inspected.}
}
\value{
a tibble; each column corresponds to the same column in dataframe and each value inside the column is 0 if the corresponding value is NA, 1 otherwise.
stops if the object passed in is not a data.frame or tibble.
}
\description{
Evaluates a dataframe for NA values.
}
\examples{
df <- data.frame(x = (c(2,3,4)), y= c(1,10,3))
col_num <- describe_na_values(df)

#> # A tibble: 2 x 3
#>       x     y
#>   <int> <int>
#> 1     1     1
#> 2     1     1
#> 3     1     1

df <- data.frame(x = (c(2,NaN,4)), y= c(1,10,3))
col_num <- describe_na_values(df)

#> # A tibble: 2 x 3
#>       x     y
#>   <int> <int>
#> 1     1     1
#> 2     0     1
#> 3     1     1

}
