% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_eda.R
\name{ebalance}
\alias{ebalance}
\title{Descriptive Statistics in the Treatment and Control Groups}
\usage{
ebalance(
  data,
  treatmentVar,
  alpha = 0.05,
  ttest = F,
  showMahalanobisDist = T,
  showpscore = T,
  showLinPscore = T,
  col.size = 120
)
}
\arguments{
\item{data}{A data frame that contains a dummy indicator
variable for the treatment and control
groups}

\item{treatmentVar}{a string with the name of the indicator
variable of the treatment}

\item{alpha}{number between 0 and 1. It is used to ...}

\item{ttest}{boolean, indicates if a t-test of difference in
means between treatment and control groups is
computed}

\item{showMahalanobisDist}{boolean, indicates if the Mahalanobis
distance is computed}

\item{showpscore}{boolean, indicates if the pscore is computed}

\item{showLinPscore}{...}

\item{col.size}{integer, size of the columns in the result
printed in the screen}
}
\value{
The function returns a list with named elements. The
        element \code{balance} constains a table with the
        statistics of all numerical variables by treatment
        and control groups. The element \code{pscore} contains
        the pscore of all observations. The element \code{lpscore}
        of the list is the logarithm of pscore.
}
\description{
This function generates a table with descriptive statistics of
numerical variables by group (Treatment versus Control groups)
}
\details{
The statistics computed here are recomended in
\itemize{
 \item Imbens, G. W., & Rubin, D. B. (2015). Causal inference in statistics, social, and biomedical sciences: an introduction. : Cambridge University Press.
}
}
\examples{
data(edar_survey)

ebalance(edar_survey , treatmentVar='treat')

# or with paper
# edar_survey  \%>\% ebalance(., treatmentVar='treat')

}
