% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar-package.R
\docType{package}
\name{edar-package}
\alias{edar}
\alias{edar-package}
\title{edar: Convenient Functions for Exploratory Data Analysis}
\description{
Convenient functions for common tasks in exploratory data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://soutomas.github.io/edar/}
  \item Report bugs at \url{https://github.com/soutomas/edar/issues}
}

}
\author{
\strong{Maintainer}: Tomas Sou \email{tomas.sou@carexer.com} (\href{https://orcid.org/0000-0002-7570-5545}{ORCID})

}
\keyword{internal}
