% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_doc_data.R
\docType{package}
\name{edar}
\alias{edar}
\alias{edar-package}
\title{edar: A package for quick Exploratory Data Analysis in R}
\description{
The package provide functions that summarize the data set,
generate tables and plots with subset of the variables,
among other features.
}
\section{Univariate EDA}{
NA
}

\section{Bivariate EDA}{
NA
}

