% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_eda.R
\name{edar_bundle_cat}
\alias{edar_bundle_cat}
\title{Smart univariate summary of the categorical variables}
\usage{
edar_bundle_cat(data, vars = NULL, print.summary = NULL,
  weight = NULL, digits = 4, width.size = 150)
}
\arguments{
\item{data}{a data frame}

\item{vars}{a vector with the name of the variables to display.
If none is provided (vars=NULL), it summarizes all
the categorical variables}

\item{print.summary}{boolean, if \code{TRUE} the function will print a summary}

\item{weight}{vector with the weights}

\item{digits}{integer, the number of digits to print}

\item{width.size}{numeric, size of the columns in the screen to print the results}

\item{print}{a string with the type of summary to print in the
screen (see \code{value} below for the complete
description of the return). Values can be one of
'prop' (display the proportions), 'count' (display
the counts), or 'prop_count' (display both the proportion
and counts).}
}
\value{
It retuns a tibble in which the columns are lists. The columns
        of the tibble are \code{var.idx}, which is a list with the idx
        of the variables, \code{vars}, which is a list with the
        variable names bundled together. Finally \code{table.p},
        \code{table.n}, and \code{table.pn} are the tables with,
        respectively, the proportions, counts, and both proportions
        and counts for the bundled variables
}
\description{
This function summarizes the distribution of the non-numerical
variables. It bundles together all the variables that have the same
categories and displays their summary in a single table.
}
\examples{
data(edar_survey)
edar_bundle_cat(edar_survey)
tb = edar_bundle_cat(edar_survey )
tb
tb$var.idx
tb$table.n
tb$table.p
tb$table.pn

# with pipe:
# edar_survey \%>\% edar_bundle_cat()

}
