% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_eda.R
\name{edar_describe}
\alias{edar_describe}
\title{Describe the variables of a data frame}
\usage{
edar_describe(data, weight = NULL, minMax = T, digits = 2, maxVars = 50,
  col.size = 120)
}
\arguments{
\item{data}{a data frame}

\item{weight}{a vector with the weights}

\item{minMax}{boolean, display the minimum and maximum in the summary
when minMax=T}

\item{digits}{the number of digits to display}

\item{maxVars}{integer, the maximum number of variables in the data to summarize.
It accepts any number, by it may result in slow computaiton of the summary
depending on the size of the data set}

\item{col.size}{integer, the size of the column to display the summary}
}
\value{
It returns a list of tibbles. The first contains the summary of the
        categorical variables, the second the summary of the numerical ones.
}
\description{
This function generate two tables. One summarizes the non-numerical
variables, the other the numerical ones.
}
\examples{
data(edar_survey)
edar_survey \%>\% edar_describe()

tb = edar_survey \%>\% edar_describe()
tb$Cat
tb$Num
}
