% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_regression.R
\name{edar_get_fitted}
\alias{edar_get_fitted}
\title{Get fitted values}
\usage{
edar_get_fitted(data, smry, formula, newdata = NULL, x = NULL, n = NULL,
  cat.values)
}
\arguments{
\item{data}{data frame containing the data used to fit the model.}

\item{smry}{a data frame with the summary of the model containing at least two columns named \code{term} and \code{estimate}. The column \code{estimate} must contains the point estimate. The columns \code{term} must match the name of the covariate associated the estimates as produced by \code{attr(terms(formula), 'term.labels')}, and \code{formula} must also be provided as a parameter.}

\item{formula}{an object of the class \code{formula} as used in the \code{lm} function. It must be the formula used to fit the model.}

\item{newdata}{either \code{NULL} (default) or a data frame with values to be used to produce the fitted values. If \code{NULL}, the new data is generated by the function \code{edar_get_new_data}}

\item{x}{string with the name of the numeric variable that will vary within the range of its values in the original data set. All other numeric covariates will be set to their average value}

\item{n}{integer, size of the new data set}

\item{cat.values}{named list of string vectors. The name of each element of the list (the string vectors) must match variable names in the data. The element of the string vectors must be strings with the name of the categories to use in the new data set. To generate the new data set, the numeric columns will be set to their mean value, except the column specified in the parameter \code{x}. The categorical values are set to their first category or the first category in alphabetic order. One can set the categorical variables to different values or use more than one category by setting this parameter \code{cat.values} as desired. For instance, suppose there is a categorical variable in the data set named education, taking the values of \code{High} or \code{Low}. If \code{cat.values=NULL}, the new data set returned will be fixed at \code{education=High}. One can use \code{education="Low"} by setting \code{cat.value=list(eductation="low")}. One can have both levels of education returned by setting \code{cat.value=list(eductation=c("low", "high"))}. See more examples in the documentation below.}
}
\description{
This function returns fitted values given model summary
}
