% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_report.R
\name{edar_report}
\alias{edar_report}
\title{Create edar report}
\usage{
edar_report(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
A summary table.
}
\description{
Generate the edar report for the given data.
}
\examples{
data(mtcars)
edar_report(mtcars)
}
