% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_survey.R
\name{epower}
\alias{epower}
\title{Compute power of a test}
\usage{
epower(
  mu1 = NULL,
  mu2 = NULL,
  power_ideal = 0.8,
  n.current = NULL,
  n1.current = NULL,
  n2.current = NULL,
  title = NULL,
  lines = T
)
}
\arguments{
\item{mu1}{a numeric vector with expected observed proportion of the first group. It must be between 0 and 1}

\item{mu2}{a number between 0 and 1 with expected observed proportion of the second group.}

\item{power_ideal}{the desired power of the test of difference between proportions}

\item{n.current}{the current/planned sample size}

\item{n1.current}{the current/planned sample size for the first group}

\item{n2.current}{the current/planned sample size for the second group}

\item{title}{string, title of the plot}

\item{lines}{boolean, if \code{TRUE} the plot generated by the function will display lines at current and ideal values of the power of the test}
}
\value{
It returns a table with the power and sample size as well as a plot with the information
}
\description{
This function compute power of a test or ideal sample size given desired power of a test
}
\examples{

epower(c(.2),.4, .8, 10)
epower(c(.2,.5),.4, .8, 10,30)

}
