% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_regression.R
\name{epredict}
\alias{epredict}
\title{Predicted values of regression}
\usage{
epredict(object, newdata)
}
\arguments{
\item{object}{an output of \code{lm}}

\item{newdata}{the new data frame to compute the predicted values}
}
\description{
Get the predicted values, includint standard errors and confidence intervals
}
