% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ft_def}
\alias{ft_def}
\title{flextable default}
\usage{
ft_def(font = "Calibri Light", fsize = 10, pad = 3)
}
\arguments{
\item{font}{<chr> See `font.family` in [flextable::set_flextable_defaults()]}

\item{fsize}{<int> Font size (in point)}

\item{pad}{<int> Padding space around}
}
\value{
A list containing previous default values
}
\description{
Sugar function to set flextable defaults.
}
\examples{
\dontrun{
ft_def()
}
}
\seealso{
[flextable::set_flextable_defaults()]
}
