% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{generate_report}
\alias{generate_report}
\title{This function generates an EDA report by plotting graphs and tables for the numeric variables, categorical variables, NA values and correlation in a dataframe}
\usage{
generate_report(dataframe, cat_vars, num_vars)
}
\arguments{
\item{dataframe}{tbl. The dataframe to be inspected.}

\item{cat_vars}{vector of character strings of the names of the categorical variables.}

\item{num_vars}{vector of character strings of the names of the numeric variables.}
}
\value{
It returns True on successful execution else returns False
}
\description{
This function generates an EDA report by plotting graphs and tables for the numeric variables, categorical variables, NA values and correlation in a dataframe
}
\examples{
X <- dplyr::tibble(type = c('Car','Bus','Car'), height = c(10,20,30))
cat_vars <- c('type')
num_vars <- c('height')
generate_report(X, cat_vars, num_vars)

}
