% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_all.R
\name{geo_cv}
\alias{geo_cv}
\title{Geometric coefficient of variation}
\usage{
geo_cv(x)
}
\arguments{
\item{x}{\verb{<num>} A vector of values.}
}
\value{
Geometric coefficient of variation
}
\description{
Compute geometric coefficient of variation (GCV)
}
\examples{
geo_cv(rlnorm(10))
}
