% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_all.R
\name{geo_mean}
\alias{geo_mean}
\title{Geometric mean}
\usage{
geo_mean(x)
}
\arguments{
\item{x}{\verb{<num>} A vector of values.}
}
\value{
Geometric mean.
}
\description{
Compute geometric mean.
}
\examples{
geo_mean(rlnorm(10))
}
