% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_all.R
\name{geo_sd}
\alias{geo_sd}
\title{Geometric standard deviation}
\usage{
geo_sd(x)
}
\arguments{
\item{x}{\verb{<num>} A vector of values.}
}
\value{
Geometric standard deviation
}
\description{
Compute geometric standard deviation (GSD)
}
\examples{
geo_sd(rlnorm(10))
}
