% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_all.R
\name{geoMean}
\alias{geoMean}
\title{Geometric mean}
\usage{
geoMean(x, ...)
}
\arguments{
\item{x}{\verb{<num>} A vector of values.}

\item{...}{Other augments to pass to \code{\link[=mean]{mean()}}.}
}
\value{
Geometric mean.
}
\description{
Compute geometric mean.
}
\examples{
geoMean(rlnorm(10))
}
