% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_all.R
\name{geoSD}
\alias{geoSD}
\title{Geometric standard deviation}
\usage{
geoSD(x, ...)
}
\arguments{
\item{x}{\verb{<num>} A vector of values.}

\item{...}{Other augments to pass to \code{\link[=sd]{sd()}}.}
}
\value{
Geometric standard deviation
}
\description{
Compute geometric standard deviation
}
\examples{
geoSD(rlnorm(10))
}
