% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ggcov_box}
\alias{ggcov_box}
\title{Box plot wrapper}
\usage{
ggcov_box(d, cont, cats = NULL, ...)
}
\arguments{
\item{d}{\verb{<dfr>} A data frame}

\item{cont}{\verb{<var>} A continuous variable to plot as unquoted name}

\item{cats}{\verb{<chr>} Optional. A character vector of selected discrete variables}

\item{...}{List of arguments to pass to \link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot}}
}
\value{
A ggplot object of a box plot
}
\description{
Sugar function to generate box plots for a continuous variable by discrete variables.
Orientation will follow the axis of the discrete variables.
}
\examples{
iris |> ggcov_box(Sepal.Length)
d <- mtcars |> dplyr::mutate(cyl=factor(cyl),gear=factor(gear),vs=factor(vs))
d |> ggcov_box(mpg)
d |> ggcov_box(mpg,c("cyl","vs"))
}
