% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_eda.R
\name{gge_barplot2}
\alias{gge_barplot2}
\title{Barplot}
\usage{
gge_barplot2(data, variable, group = NULL, title = NULL, subtitle = NULL,
  legend.position = "top")
}
\arguments{
\item{data}{a data frame}

\item{variable}{string with the name of the categorical variable that is going to be used to create the plot}

\item{group}{a string with the name of the varible(s) to group the plots}

\item{title}{either \code{NULL} or a string with the title of the plot}

\item{subtitle}{either \code{NULL} or a string with the subtitle of the plot}

\item{legend.position}{a string (\code{top}, \code{bottom}, \code{left} (Default)), \code{right}}
}
\description{
This function generate a barplot
}
