% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_eda.R
\name{gge_contour}
\alias{gge_contour}
\title{Contour plot}
\usage{
gge_contour(df, x, y, points = FALSE, group = NULL, palette = "Blues",
  hist.fill = "lightsteelblue2", hist.col = "white", xlim = NULL,
  xlab = NULL, ylab = NULL)
}
\arguments{
\item{df}{a data frame}

\item{x}{string with the name of the variable to plot in the x-axis}

\item{y}{string with the name of the variable to plot in the y-axis}

\item{points}{boolean, if \code{TRUE} displays the points}

\item{group}{a string with the categories to group the points}

\item{palette}{a string with the name of a color brewer continuous palette}

\item{hist.fill}{a string with the colour to fill the histogram bars}

\item{hist.col}{a string with the color of the border of the histogram bars}

\item{xlim}{two dimensional numeric vectors with the limits of the x-axis}

\item{xlab}{a string to be displayed in the x-axis}

\item{ylab}{a string to be displayed in the y-axis}
}
\description{
Contour plot
}
