% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_eda.R
\name{gge_density}
\alias{gge_density}
\title{Plot densities}
\usage{
gge_density(
  df,
  group = NULL,
  weight = NULL,
  mean = TRUE,
  median = FALSE,
  conf.int = TRUE,
  quantile = FALSE
)
}
\arguments{
\item{df}{a data frame}

\item{group}{a string with the name of the varible(s) to group the plots}

\item{weight}{a string with the name of the varible containing the weights. They are used to compute the summary stattistics}

\item{mean}{boolean, if \code{TRUE}, displays a line with the averave value}

\item{median}{boolean, if \code{TRUE}, displays a line with the median value}

\item{conf.int}{boolean, if \code{TRUE}, displays lines with the 95\% confidence interval (it is meaningful for variables that normally distributed)}

\item{quantile}{boolean, if \code{TRUE}, displays lines with the 25\% and 75\% quantiles}
}
\description{
Plot densities of all numerical variables in the data frame
}
