% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_eda.R
\name{gge_describe}
\alias{gge_describe}
\title{Descriptive Plots}
\usage{
gge_describe(df, group = NULL, weight = NULL, mean = TRUE,
  median = FALSE, conf.int = TRUE, quantile = FALSE, n.plots = 16,
  common.legend = TRUE, hist = FALSE, legend.position = "top")
}
\arguments{
\item{df}{a data frame}

\item{group}{a string with the name of the varible(s) to group the plots}

\item{weight}{a string with the name of the varible containing the weights. They are used to compute the summary stattistics}

\item{mean}{boolean, if \code{TRUE}, displays a line with the averave value}

\item{median}{boolean, if \code{TRUE}, displays a line with the median value}

\item{conf.int}{boolean, if \code{TRUE}, displays lines with the 95\% confidence interval (it is meaningful for variables that normally distributed)}

\item{quantile}{boolean, if \code{TRUE}, displays lines with the 25\% and 75\% quantiles}

\item{n.plots}{integer, the number of plots to display in the grid}

\item{common.legend}{boolean, if \code{TRUE} the plots uses a single legend}

\item{hist}{boolean, if \code{TRUE}, histograms are displayed instead of densities for the numerical variables}

\item{legend.position}{a string (\code{top}, \code{bottom}, \code{left} (Default)), \code{right}}
}
\description{
This function generate plots with all the variables in the data frame individually
}
