% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_eda.R
\name{gge_fdensity}
\alias{gge_fdensity}
\title{Plot densities by group and facet, with Kolmogorov-Smirnov statistics}
\usage{
gge_fdensity(df, value, group, facet, scale = "fixed")
}
\arguments{
\item{df}{a data frame}

\item{value}{a string with the name of the numerical value to plot the density}

\item{group}{a string with the name of the variable to group the densities}

\item{facet}{a string with the name of the variable to create the grid of plots}

\item{scale}{see \code{\link[ggplot2]{facet_wrap}}}
}
\description{
Plot densities by group and facet, with Kolmogorov-Smirnov statistics
}
\details{
All the variables in the data frame will be plotted
}
