% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gg.R
\name{ggout}
\alias{ggout}
\title{Save ggplot with output path}
\usage{
ggout(plt, fpath, lab = "", omit = "", ...)
}
\arguments{
\item{plt}{A ggplot object.}

\item{fpath}{\verb{<chr>} File path to save output and add to label.}

\item{lab}{\verb{<chr>} Custom label to use instead of \code{fpath}.}

\item{omit}{\verb{<chr>} Text to omit from the label.}

\item{...}{Other arguments to pass to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}
}
\value{
The file path of the output.
}
\description{
Save ggplot with output path
}
\examples{
\dontrun{
fpath = "../output.png"
iris |> gghist() |> ggout(fpath)
}
}
\seealso{
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}
}
