% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_gg.R
\name{ggtp}
\alias{ggtp}
\title{Create time-profile plots}
\usage{
ggtp(
  d,
  x,
  y,
  id,
  ...,
  nsub = TRUE,
  logx = FALSE,
  logy = FALSE,
  alpha_point = 0.2,
  alpha_line = 0.1,
  xlab = NULL,
  ylab = NULL,
  ttl = NULL,
  sttl = NULL,
  cap = NULL
)
}
\arguments{
\item{d}{\verb{<dfr>} A data frame.}

\item{x, y}{\verb{<var>} Column for x- and y-axis.}

\item{id}{\verb{<var>} Column of subject ID for grouping.}

\item{...}{Other arguments to pass to \link[ggplot2:aes]{ggplot2::aes} for additional mapping.}

\item{nsub}{\verb{<lgl>} \code{TRUE} to show number of subjects in caption.}

\item{logx, logy}{\verb{<lgl>} \code{TRUE} to log x- and y-axis.}

\item{alpha_point}{\verb{<num>} Alpha value for \link[ggplot2:geom_point]{ggplot2::geom_point}.}

\item{alpha_line}{\verb{<num>} Alpha value for \link[ggplot2:geom_path]{ggplot2::geom_line}.}

\item{xlab, ylab}{\verb{<chr>} Labels for x- and y-axis.}

\item{ttl, sttl, cap}{\verb{<chr>} Title. Subtitle. Caption.}
}
\value{
A ggplot object.
}
\description{
Wrapper function to create plots for time profiles such as PK and PD plots.
}
\examples{
Theoph |> ggtp(x=Time,y=conc,id=Subject)
}
