% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_gg.R
\name{ggviolin}
\alias{ggviolin}
\title{Violin plot wrapper by categorical covariates}
\usage{
ggviolin(d, var, cats, ...)
}
\arguments{
\item{d}{\verb{<dfr>} A data frame.}

\item{var}{\verb{<var>} A variable to plot as unquoted name.}

\item{cats}{\verb{<var>} Optional. Categorical variables to plot as a vector of unquoted names.}

\item{...}{List of arguments to pass to \link[ggplot2:geom_violin]{ggplot2::geom_violin}.}
}
\value{
A ggplot object with violin plots.
}
\description{
Create violin plots for a chosen variable by all discrete covariates in a dataset.
Orientation will follow the axis of the discrete variables.
Numeric variables will be dropped, except the chosen variable to plot.
}
\examples{
iris |> ggviolin(Sepal.Length)
sleep |> ggviolin(extra)
sleep |> ggviolin(extra,group)
d <- mtcars |> dplyr::mutate(cyl=factor(cyl),gear=factor(gear),vs=factor(vs))
d |> ggviolin(mpg)
d |> ggviolin(mpg,c("cyl","vs"))
}
