% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_gg.R
\name{ggxy}
\alias{ggxy}
\title{XY scatter plot wrapper}
\usage{
ggxy(d, x, y, lm = TRUE, se = TRUE, cor = TRUE)
}
\arguments{
\item{d}{\verb{<dfr>} A data frame.}

\item{x}{\verb{<var>} Variable for the x-axis as unquoted name.}

\item{y}{\verb{<var>} Variable for the y-axis as unquoted name.}

\item{lm}{\verb{<lgl>} \code{TRUE} to add regression line from linear model.}

\item{se}{\verb{<lgl>} \code{TRUE} to show standard error with the regression line.}

\item{cor}{\verb{<lgl>} \code{TRUE} to show Pearson correlation coefficient with p-value.}
}
\value{
A ggplot object.
}
\description{
Create basic XY scatter plot for quick data exploration.
Default to show Pearson correlation coefficient with p-value from \code{\link[ggpubr:stat_cor]{ggpubr::stat_cor()}}.
For more complex plots, it is recommended to use \link[ggplot2:ggplot2-package]{ggplot2::ggplot2} directly.
}
\examples{
mtcars |> ggxy(wt,hp)
mtcars |> ggxy(mpg,hp,se=FALSE)
mtcars |> ggxy(mpg,hp,lm=FALSE)
mtcars |> ggxy(cyl,hp,cor=FALSE)
}
