% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_misc.R
\name{hexn}
\alias{hexn}
\title{Generate hex colour codes}
\usage{
hexn(n, show = FALSE)
}
\arguments{
\item{n}{\verb{<int>} Number of colours to output.}

\item{show}{\verb{<lgl>} \code{TRUE} to show the output colours.}
}
\value{
A vector of hex colour codes that can be used for plotting.
}
\description{
Generate a vector of hex colour codes for the desired number of colours.
Colours are generated by evenly splitting hue in the range \verb{[0,360]}
in the HCL colour space using \link[grDevices:hcl]{grDevices::hcl}.
The output is meant to generate the default colours used in \link[ggplot2:ggplot2-package]{ggplot2::ggplot2}.
}
\examples{
hexn(6,FALSE)
hexn(4,TRUE)
}
