% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_util.R
\name{interpolate_blinks}
\alias{interpolate_blinks}
\title{Interpolate blinks per trial}
\usage{
interpolate_blinks(obj, time_new, thin = 100)
}
\description{
First, data is thinned by taking every `thin`-th value.
Then blinks are detected by high values of the derivative and 
interpolated linearly.
}

