% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_survey.R
\name{isco88toEGP}
\alias{isco88toEGP}
\title{Recode occupation to DGP class scheme}
\usage{
isco88toEGP(isco88, n.employees = NULL, self.employed = NULL,
  collapse = FALSE)
}
\arguments{
\item{isco88}{numeric vector with the isco88 4 digit code.}

\item{n.employees}{a vector containing the number of employees.}

\item{self.employed}{a vector containing the value 1 whenever the person is self-employed}

\item{collapse}{boolean, if TRUE, returns a 7 categories class scheme. Otherwise (default), it returns the 11 default categories.}
}
\description{
The function creates a factor with the EGP class scheme
}
