% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_occupation.R
\name{isco88toESeC}
\alias{isco88toESeC}
\title{Compute ESeC}
\usage{
isco88toESeC(
  isco88,
  employed = NULL,
  unemployed = NULL,
  self.employed = NULL,
  supervisor = NULL,
  n.employees = NULL,
  n.cat = 10,
  isco88.armed.forces = NULL
)
}
\arguments{
\item{isco88}{a vector with the isco88 codes}

\item{employed}{a vector containing the value 1 whenever the person is employed}

\item{unemployed}{a vector containing the value 1 whenever the person is unemployed}

\item{self.employed}{a vector containing the value 1 whenever the person is self-employed}

\item{supervisor}{a vector containing the value 1 whenever the person is a supervisor}

\item{n.employees}{a vector containing the number of employees.}

\item{n.cat}{either one of the following integers: 10 (default), 6 ,5, or 3 indicating the number of categories to be used inthe ESEeC scheme}

\item{isco88.armed.forces}{numeric vector with the code for armed forces (usually isco88 major code 0)}
}
\value{
It returns a list with two elements, the ESec code and its label
}
\description{
This function returns the ESeC (European Socio-economic classification) scheme
}
