% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_all.R
\name{kb}
\alias{kb}
\title{kable wrapper}
\usage{
kb(d, fnote = NULL, cap = NULL, sig = 3, dig = NULL, src = 0, omit = "")
}
\arguments{
\item{d}{\verb{<dfr>} A data frame.}

\item{fnote}{\verb{<chr>} Footnote.}

\item{cap}{\verb{<chr>} Caption.}

\item{sig}{\verb{<int>} Number of significant digits to compute.}

\item{dig}{\verb{<int>} Number of decimal places to display.}

\item{src}{\verb{<int>} Either 1 or 2 to add source label over 1 or 2 lines.}

\item{omit}{\verb{<chr>} Text to omit from the source label.}
}
\value{
A kable object.
}
\description{
Sugar function for default kable output.
}
\examples{
mtcars |> head() |> kb()
mtcars |> head() |> kb(src=1)
mtcars |> head() |> kb("Footnote")
mtcars |> head() |> kb("Footnote",src=1)
mtcars |> head() |> kb(sig=2,dig=1)
}
