% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{kbl2}
\alias{kbl2}
\title{kable wrapper}
\usage{
kbl2(d, fnote = NULL, cap = NULL, sig = 3, dig = 3)
}
\arguments{
\item{d}{A data frame}

\item{fnote}{[optional] <chr> Footnote}

\item{cap}{[optional] <chr> Caption}

\item{sig}{<int> Number of significant digits to compute}

\item{dig}{<int> Number of decimal places to display}
}
\value{
A kable object
}
\description{
Sugar function for default kable output.
}
\examples{
iris |> kbl2()
mtcars |> kbl2()
mtcars |> kbl2(sig=2,dig=1)
}
