% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitEDA.R
\name{knitEDA}
\alias{knitEDA}
\alias{knitEDA.common_list}
\alias{knitEDA.default}
\alias{knitEDA.plot_list}
\alias{knitEDA.vect_list}
\title{knitEDA}
\usage{
knitEDA(x, ...)

\method{knitEDA}{common_list}(x, ...)

\method{knitEDA}{plot_list}(x, ...)

\method{knitEDA}{vect_list}(x, ...)

\method{knitEDA}{default}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{
prints stuff out
}
\description{
knitEDA
}
\section{Methods (by class)}{
\itemize{
\item \code{common_list}: common_list method

\item \code{plot_list}: plot_list method

\item \code{vect_list}: vect_list method

\item \code{default}: default method (does not print and warns)
}}
\examples{
uhg an example
}

