% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{label_tz}
\alias{label_tz}
\title{Generate time stamp label}
\usage{
label_tz(omit = "")
}
\arguments{
\item{omit}{[optional] <chr> Text to omit from the label}
}
\value{
A label with time stamp
}
\description{
Generate a time stamp label of the current time.
}
\examples{
label_tz()
}
