% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load_data
Function for loading data from multiple .asc files into R.}
\usage{
load_data(datadir, savedir = "", prefix = "", trial_seperators = NULL,
  with_images = FALSE, discard_oot_data = FALSE, ncpus = 1, ...)
}
\arguments{
\item{datadir}{[character vector] with the path to the folder where the data is located. If of length
greater than 1, it is assumed to be a vector of paths to files that are to be read in.}
}
\description{
load_data
Function for loading data from multiple .asc files into R.
}

