% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movies}
\alias{movies}
\title{Data set with information about movies.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 651 rows and 32 columns.
}
\source{
\url{http://www2.stat.duke.edu/~mc301/data/movies.html}
}
\usage{
data(movies)
}
\description{
Data set with information about movies.
}
\keyword{datasets}
