% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{plot.edar_balance}
\alias{plot.edar_balance}
\title{Plot}
\usage{
\method{plot}{edar_balance}(x, ...)
}
\arguments{
\item{x}{an output of the function \code{balance}}

\item{...}{ignored}
}
\description{
Generic method to plot the output of a \code{balance}
}
