% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.edar_data}
\alias{plot.edar_data}
\title{Plotting functions}
\usage{
\method{plot}{edar_data}(obj, what = c("gaze", "ps"), subs = 1:.N,
  limit_screen = TRUE, ...)
}
\description{
Plotting functions
}

