% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_regression.R
\name{plot_diagnostic}
\alias{plot_diagnostic}
\title{Generate diagnostics for lm models}
\usage{
plot_diagnostic(model)
}
\arguments{
\item{model}{a \code{lm} object}
}
\description{
Generate diagnostics for lm models
}
