% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_eda.R
\name{plot_edescribe}
\alias{plot_edescribe}
\title{Descriptive Plots}
\usage{
plot_edescribe(data, qline = T, quantiles = c(0.025, 0.975), ...)
}
\arguments{
\item{data}{a data frame}

\item{qline}{boolean, if \code{TRUE} the plots will display lines in the quantiles}

\item{quantiles}{vector with values between 0 and 1 indicating the quantiles to display}
}
\description{
This function generate plots with all the variables in the data frame individually. It is a faster version of gge_describe
}
\details{
All the variables in the data frame will be plotted
}
