% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_eda.R
\name{plot_heat}
\alias{plot_heat}
\title{Heat plot}
\usage{
plot_heat(
  x = "",
  y = "",
  xlab = "x",
  ylab = "y",
  title = "",
  text = "",
  axis = TRUE
)
}
\arguments{
\item{x, y}{numeric vector with values to plot}

\item{xlab}{string to be displayed in the x-axis}

\item{ylab}{string to be displayed in the y-axis}

\item{title}{string with the title of the plot}

\item{text}{text annotation}

\item{axis}{boolean, if \code{TRUE} the axis are displayed}
}
\description{
Heat plot
}
