% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot histogram}
\usage{
plot_histogram(data, column)
}
\arguments{
\item{data}{data.frame}

\item{column}{string}
}
\value{
plots
}
\description{
Plot histogram
}
