% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_stat_dist.R
\name{plot_mvgaussian}
\alias{plot_mvgaussian}
\title{Plot bivariate gaussian distribution}
\usage{
plot_mvgaussian(mu = c(0, 0), Sigma = diag(2))
}
\arguments{
\item{mu}{a two-dimensional numerical vector with the mean parameters of the bivariate gaussian distribution}

\item{Sigma}{a 2x2 covariance matrix}
}
\description{
This functions generates a contour plot of a bivariate gaussian distribution
}
\examples{

}
