% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_stat_dist.R
\name{plot_mvgaussian3D}
\alias{plot_mvgaussian3D}
\title{3D gaussian distribution}
\usage{
plot_mvgaussian3D(mu = c(0, 0), Sigma = diag(2))
}
\arguments{
\item{mu}{a two-dimensional numerical vector with the mean parameters of the bivariate gaussian distribution}

\item{Sigma}{a 2x2 covariance matrix}
}
\description{
3D gaussian distribution
}
