% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_eda.R
\name{print.edar_balance}
\alias{print.edar_balance}
\title{Print}
\usage{
\method{print}{edar_balance}(x, digits = 4, ...)
}
\arguments{
\item{x}{an output of the function \code{ebalance}}

\item{digits}{integer, number of digits to display}

\item{...}{ingored}
}
\description{
Generic method to print the output of a \code{ebalance}
}
