% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printSingleVarStats.R
\name{printSingleVarStats}
\alias{printSingleVarStats}
\title{Create a full knit of a dataframe}
\usage{
printSingleVarStats(single_var_stat_object)
}
\arguments{
\item{x}{a dataframe}
}
\description{
Create a full knit of a dataframe
}

