% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ascii.R
\name{read_ascii}
\alias{read_ascii}
\title{read_ascii 
Function for loading raw data from a single .asc file into R.}
\usage{
read_ascii(path_ascii, save_rdata = "", number_trials = NULL,
  trial_seperators = NULL, ...)
}
\arguments{
\item{trial_seperators}{[list] with entries 'trial_start' and 'trial_end'. Either these entries contain 
a character vector of length 1 with the message delineating the trials, or a character with one message
per trial. For example, the whole time series of the experiment is cut between the first entries of 
`trial_start` and `trial_end` to obtain the first trial.
`trial_end` can also be [numeric] and indicating the length trials.}
}
\description{
read_ascii 
Function for loading raw data from a single .asc file into R.
}

