% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_survey.R
\name{recode.likert5}
\alias{recode.likert5}
\title{Recode variable that uses likert scale}
\usage{
recode.likert5(df, vars.to.recode, invert = FALSE, new.levels = NULL)
}
\arguments{
\item{df}{a data.frame}

\item{vars.to.recode}{a string vector with the names of the variables to recode}

\item{invert}{boolean, indicates if the values of the new variable should be inverted}

\item{new.levels}{a string vector with the name of the new categories}
}
\value{
The function returns the data frame with 6 new variables named <vars.to.recode>5 (numerical variable with 5 levels from 1 to 5),<vars.to.recode>5c (factor with 5 levels), <vars.to.recode>3 (numerical variable with 3 levels, -1, 0, and 1), <vars.to.recode>3c (3 levels categorical variable), <vars.to.recode>2 (2 levels numerical variable), <vars.to.recode>2c (two levels categorical variable).
}
\description{
This function can be used to quickly recode variables that use 5 points likert scale
}
