% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_eda.R
\name{recode}
\alias{recode}
\title{Recode labelled columns}
\usage{
recode(
  df,
  var,
  newvar = var,
  newvar_label = NULL,
  missing = NULL,
  recode,
  sep = " / "
)
}
\arguments{
\item{df}{a data frame}

\item{var}{a string with the name of the variable in the data frame
to be recoded}

\item{newvar}{a string with the name of the new variable that will
receive the recoded columns. By default, it uses the
original variable (var)}

\item{newvar_label}{string with the label for the new variable}

\item{recode}{a named vector. The names are the code of the labels in the
old variable. The values are the new values}

\item{sep}{a string to use to separate labels when joining values
during the recoding}
}
\value{
It returns a tibble data frame with recoded column
}
\description{
This function recode labelled columns
}
\examples{


}
