% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_all.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{mutate}
\alias{filter}
\alias{select}
\alias{where}
\alias{across}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{across}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:reexports]{where}}}
}}

