% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_utils.R
\name{reorder_within}
\alias{reorder_within}
\title{Reorder labels}
\usage{
reorder_within(x, by, within, fun = mean, sep = "___", ...)
}
\arguments{
\item{x}{the name of the variable that need to be ordered}

\item{by}{the name of the variable that will be used to order \code{x}}

\item{within}{the name of the variable used to create the facets}

\item{fun}{(not used)}

\item{sep}{default variable to manipulate the labels. Default "__"}

\item{...}{additional parameter...}
}
\description{
This function reorder labels in a ggplot2 barplot figure with facets (see examples)
}
