% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_eda.R
\name{rescale.axis}
\alias{rescale.axis}
\title{Rescale values}
\usage{
rescale.axis(y.reference, y.to.rescale)
}
\arguments{
\item{y.reference}{values to use as reference to rescale}

\item{y.to.rescale}{values to rescale into y.left range}
}
\description{
Rescale values
}
\examples{

}
