% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{roi}
\alias{add_roi}
\alias{roi}
\title{Region-of-interest (ROI)
Functions for generating and analysing ROIs.}
\usage{
roi(xy1, xy2)

add_roi(obj, roi_obj)
}
\description{
ROI object

Function for adding a ROI to the dataset, which can then be used
for analysis.
}
\section{Functions}{
\itemize{
\item \code{add_roi}: adding ROI to dataset
}}

