% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save.figure}
\alias{save.figure}
\title{Save Figure and Associated Table}
\usage{
save.figure(
  g,
  tab,
  caption,
  path,
  fn,
  width = NA,
  height = NA,
  add.to.caption = NULL
)
}
\arguments{
\item{g}{A ggplot object to save.}

\item{tab}{A data frame or table that contains the data used to create the figure.}

\item{caption}{A text caption for the figure.}

\item{path}{A directory path where the files will be saved.}

\item{fn}{The base filename (without extension) for the saved files.}

\item{width}{Numeric value for the width of the figure. Default is NA.}

\item{height}{Numeric value for the height of the figure. Default is NA.}

\item{add.to.caption}{Optional additional text to append to the figure's caption.}
}
\value{
A message indicating the completion of the saving process.
}
\description{
This function saves a figure generated by ggplot2 and its associated data table in multiple formats.
}
