% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_utils.R
\name{scale_x_reordered}
\alias{scale_x_reordered}
\title{Reorder labels in ggplot}
\usage{
scale_x_reordered(..., sep = "___")
}
\arguments{
\item{...}{additional parameter...}

\item{sep}{default variable to manipulate the labels. Default "__"}
}
\description{
Function used in conjunction with \link{reorder_within} to reorder labels based on a numerical variable in plots with facets
}
