% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/singleVarStats.R
\name{singleVarStats}
\alias{singleVarStats}
\alias{singleVarStats.Date}
\alias{singleVarStats.character}
\alias{singleVarStats.factor}
\alias{singleVarStats.numeric}
\title{Meat of the product}
\usage{
singleVarStats(x, ...)

\method{singleVarStats}{factor}(x, trim = FALSE,
  max_list = getOption("max.print"), var_name = "def_x")

\method{singleVarStats}{character}(x, trim = FALSE,
  max_list = getOption("max.print"), var_name = "x_def")

\method{singleVarStats}{numeric}(x, trim = FALSE,
  max_list = getOption("max.print"), var_name = "x_def")

\method{singleVarStats}{Date}(x, trim = FALSE,
  max_list = getOption("max.print"), var_name = "x_def")
}
\arguments{
\item{x}{}

\item{...}{}

\item{trim}{}

\item{max_list}{}

\item{var_name}{}

\item{x}{}
}
\value{
object

ok
}
\description{
Meat of the product

factor version
}
\section{Methods (by class)}{
\itemize{
\item \code{factor}: 

\item \code{character}: 

\item \code{numeric}: 

\item \code{Date}: 
}}
\examples{
nice
test
}

