% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{skim_data}
\alias{skim_data}
\title{Skim data}
\usage{
skim_data(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
data.frame
}
\description{
Return data.frame with the columns and the corresponding type.
}
