% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/skipPrintEDA.R
\name{skipPrintEDA}
\alias{skipPrintEDA}
\title{Add a skip class to an object}
\usage{
skipPrintEDA(x)
}
\arguments{
\item{x}{}
}
\value{
original item with skip_eda class attached.
}
\description{
Add a skip class to an object
}
\keyword{internal}

