% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_misc.R
\name{summ_cat}
\alias{summ_cat}
\title{Summarise categorical variables}
\usage{
summ_cat(d, ...)
}
\arguments{
\item{d}{A data frame.}

\item{...}{\verb{<var/int>} (name or index) Optional. A variable to extract.}
}
\value{
A list containing summaries for each categorical variables.
}
\description{
Summarise all categorical variables.
Numeric variables will be dropped.
}
\examples{
x <- mtcars |> dplyr::mutate(dplyr::across(c(cyl,vs,am,gear,carb),factor))
x |> summ_cat()
x |> summ_cat(1)
x |> summ_cat(cyl)
}
