% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_eda.R
\name{summarise_allc}
\alias{summarise_allc}
\title{Summarise the categorical variables of the data set}
\usage{
summarise_allc(df, group = NULL, ...)
}
\arguments{
\item{df}{data frame}

\item{group}{string vector with the names of the grouping variables (can be more than one). The summary will be computed by groups. Default \code{NULL}.}
}
\value{
It returns a tibble data frame with summaries for all categorical variables.
}
\description{
This function summarises all the categorical variables in the data set.
}
\examples{
data(starwars)

starwars  \%>\% summarise_allc(., group=NULL)
starwars  \%>\% summarise_allc(., group="gender")
starwars  \%>\% summarise_allc(., group=c("gender", "eye_color"))

}
