% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_eda.R
\name{summarise_allcbundle}
\alias{summarise_allcbundle}
\title{Summarise categorical variables in bundles}
\usage{
summarise_allcbundle(df, group = NULL)
}
\arguments{
\item{df}{data frame}

\item{group}{string vector with the names of the grouping variables (can be more than one). The summary will be computed by groups. Default \code{NULL}.}
}
\value{
It returns a tibble data frame with summaries for all categorical variables. The summaries are aggregated into single tables based on the category labels.
}
\description{
This function summarises all the categorical variables in the data set and create bundles based on the category labels
}
\examples{

data(edar_survey)
## help(edar_survey)
tab =  summarise_allcbundle(edar_survey)
tab
tab$Table[[4]]
tab$Tablep[[4]]
tab$Tablel[[4]]
tab = summarise_allcbundle( edar_survey, group="gender")
tab
tab$Table[[5]]
tab$Tablep[[5]]
tab$Tablel[[5]]

# or with pipe
# edar_survey  \%>\% summarise_allcbundle(., group="gender")

}
