% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_eda.R
\name{summarise_alln}
\alias{summarise_alln}
\title{Summarise numerical variables of the data set}
\usage{
summarise_alln(df, group = NULL, weight = NULL, spread = F)
}
\arguments{
\item{df}{data frame}

\item{group}{string vector with the names of the grouping variables (can be more than one). The summary will be computed by groups. Default \code{NULL}.}

\item{weight}{string vector with the name of the column in the data set containing the weights. Default \code{NULL}.}

\item{spread}{boolean, if \code{TRUE} the summaries for each group are spread along the columns. Default \code{FALSE}}
}
\value{
It returns a tibble data frame with summaries for all numerical variables
}
\description{
This function summarises all the numerical variables in the data set.
}
\examples{
data(starwars, package='dplyr')

summarise_alln(starwars, group=NULL, weight=NULL, spread=FALSE)
summarise_alln(starwars, group="gender", weight=NULL, spread=FALSE)
summarise_alln(starwars, group="gender", weight=NULL, spread=TRUE)

# or use with pipe
# starwars  \%>\% summarise_alln(., group="gender", weight=NULL, spread=T)
}
