% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_all.R
\name{tab2v}
\alias{tab2v}
\title{Table wrapper for two variables}
\usage{
tab2v(d, x, y)
}
\arguments{
\item{d}{\verb{<dfr>} A data frame.}

\item{x, y}{\verb{<var>} Variables to tabulate.}
}
\value{
A tabyl object.
}
\description{
Wrapper function to produce summary tables for two variables.
}
\examples{
# example code
mtcars |> tab2v(vs,cyl)
mtcars |> tab2v(vs,am)
mtcars |> tab2v(vs,am)
mtcars |> tab2v(vs,gear)
}
