% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{tbl}
\alias{tbl}
\title{Table Function}
\usage{
tbl(
  data,
  v1,
  v2,
  groups = NULL,
  margin = "all",
  stat = "both",
  total = c("row", "col"),
  use.labels = T,
  digits = 1,
  na.label = "NA"
)
}
\arguments{
\item{data}{A data frame containing the data to be analyzed.}

\item{v1}{The first variable for the contingency table. No need to be quoted.}

\item{v2}{The second variable for the contingency table. No need to be quoted.}

\item{groups}{Optional grouping variable of vector of variables. No need to be quoted.}

\item{margin}{Specifies the margin for proportions ('all', 'row', or 'col').}

\item{stat}{The type of statistic to calculate: "n" for counts, "perc" for percentages, or "both" for both.}

\item{total}{Specifies whether to include totals for rows or columns ("row", "col").}

\item{use.labels}{Logical; should labels from the data be used?}

\item{digits}{The number of digits to display for percentages.}

\item{na.label}{The label to use for NA values.}
}
\value{
A printed table of counts and/or proportions as specified, with totals where applicable.
}
\description{
This function creates a contingency table from two specified variables in a dataset, with optional grouping, statistical calculations, and formatting.
}
