% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_mediate}
\alias{tidy_mediate}
\title{Tidy mediate output}
\usage{
tidy_mediate(x)
}
\arguments{
\item{x}{output of mediate function}
}
\description{
This functions creates a tibble summary with output of mediate function (mediation package)
}
