% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_summary_stat.R
\name{tidye_mediate}
\alias{tidye_mediate}
\title{Tidy mediate output}
\usage{
tidye_mediate(x)
}
\arguments{
\item{x}{output of mediate function}
}
\description{
This functions creates a tibble summary with output of mediate function (mediation package)
}
