% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_regression.R
\name{tidyehc}
\alias{tidyehc}
\title{Tidy linear model summary with robust std error}
\usage{
tidyehc(model, type = c("hc3", "hc0", "hc1", "hc2", "hc4"), digits = 4)
}
\arguments{
\item{model}{a \code{lm} of \code{glm} model}

\item{type}{a string with the method to compute the standard errors (see \code{\link[car]{hccm}})}

\item{digits}{integer, the number of significant digitis to use}
}
\description{
This function returns a tidy data.frame with output summaries for \code{lm} and \code{glm} models. Both robust standard errors (homoskedasticity corrected) and default standard errors are returned
}
