% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{transf}
\alias{transf}
\title{Function for transforming a point cloud.}
\usage{
transf(pointPath, from = list(bottomleft = c(-1, -1), topright = c(1, 1)),
  to = list(bottomleft = c(0, 0), topright = c(1, 1)), shift = c(0, 0),
  flip = FALSE)
}
\description{
Rescaling from some coordinate system, defined by argument 'from' to another
given by 'to'. If 'shift' is given, the rescaled cloud is moved into the center
of a rectangle defined by shift.
}

