% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/commonSingVar.R
\name{commonSingVar}
\alias{commonSingVar}
\title{Common statistics that are applied to all vector classes.}
\usage{
commonSingVar(x, trim = FALSE, max_list = getOption("max.print"),
  var_name = "def_x")
}
\arguments{
\item{x}{A vector to perform operations on.}

\item{trim}{Should the output be trimmed for better readability?}

\item{max_list}{the amount of members to show for tables.}

\item{var_name}{Currently not used, may be used for labeling later on.}
}
\value{
a list with the elements outlined above
}
\description{
\code{commonSingVar()} returns a list with the following members described in
detail below: length, naCnt, naPerc, unique_length, uniquePerc, final,
full_table. It is mainly used in the \link{singleVarStats} function.
}
\details{
\code{length}: the number of elements of the vector.

\code{naCnt}: number of \code{NA}s in the vector.

\code{naPerc}: naCnt/length.

\code{unique_length}: number of unique values

\code{uniqiePerc}: unique_length/length

\code{final}: full vector

\code{full_table}: a table with the count and \link{prop.table} for each value,
sorted by count
}
\examples{
out <- commonSingVar(edaR_test$state.name)
str(out)

## List of 7
## $ length       : int 50
## $ naCnt        : int 6
## $ naPerc       : num 0.12
## $ unique_length: int 45
## $ uniquePerc   : num 0.9
## $ final        : chr [1:50] "Alabama" NA "Arizona" NA ...
## $ full_table   : num [1:44, 1:2] 1 1 1 1 1 1 1 1 1 1 ...
## ..- attr(*, "dimnames")=List of 2
## .. ..$ : chr [1:44] "Alabama" "Arizona" "California" "Colorado" ...
## .. ..$ : chr [1:2] "ct" "prop"
## - attr(*, "class")= chr [1:2] "common_list" "list"
}

