% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/commonSingVar.R
\name{commonSingVar}
\alias{commonSingVar}
\title{Common statistics that are applied to all vector classes.}
\usage{
commonSingVar(x, trim = FALSE, max_list = getOption("max.print"),
  var_name = "def_x")
}
\arguments{
\item{x}{A vector.}

\item{trim}{Should the output be trimmed for better readability?}

\item{max_list}{the amount of members to show for tables.}

\item{var_name}{Currently not used, may be used for labeling later on.}
}
\value{
a list with the following elements:
 length
 full_table (proportions and counts of members of x)
 NA Count
 NA Percentage
 Length of Unique Values
 Final
}
\description{
Common statistics that are applied to all vector classes.
}

