% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar-package.R
\docType{package}
\name{edar-package}
\alias{edar}
\alias{edar-package}
\title{edar: Convenient Functions for Exploratory Data Analysis}
\description{
A collection of convenient functions to facilitate common tasks in exploratory data analysis. Some common tasks include generating summary tables of variables, displaying tables as a 'flextable' or a 'kable' and showing distributions of variables using 'ggplot2'. Labels stating the source file with run time can be easily generated for annotation in tables and plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://soutomas.github.io/edar/}
  \item \url{https://github.com/soutomas/edar/}
  \item Report bugs at \url{https://github.com/soutomas/edar/issues}
}

}
\author{
\strong{Maintainer}: Tomas Sou \email{tomas.sou@carexer.com} (\href{https://orcid.org/0000-0002-7570-5545}{ORCID})

}
\keyword{internal}
