% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar-package.R
\docType{package}
\name{edar-package}
\alias{edar}
\alias{edar-package}
\title{edar: Convenient Functions for Exploratory Data Analysis}
\description{
Convenient functions for common tasks in exploratory data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/soutomas/edar}
  \item Report bugs at \url{https://github.com/soutomas/edar/issues}
}

}
\author{
\strong{Maintainer}: Tomas Sou \email{tomas.sou@carexer.com} [copyright holder]

}
\keyword{internal}
