% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{_PACKAGE}
\name{edar}
\alias{edar}
\title{edar: A Package for Code-Efficient Exploratory Data Analysis in R}
\description{
The `edar` package provides functions that allow for efficient exploratory
data analysis with minimal lines of code. It includes various functions for
checking the balance of covariates between control and treatment groups, 
analyzing the outputs of model estimations, creating summary tables and 
plots, and conducting robustness checks of model results (e.g., multiple 
imputation, post-stratification). These tasks are frequently performed 
by quantitative researchers, and this package facilitates their completion 
quickly and with minimal coding effort.
}
\details{
For examples of workflows, see the vignette: `vignette("edar")`
}
