% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_regression.R
\name{edar_get_new_data}
\alias{edar_get_new_data}
\title{short title}
\usage{
edar_get_new_data(data, n, x, group1.label = NULL, group1.value = NULL,
  group2.label = NULL, group2.value = NULL, group3.label = NULL,
  group3.value = NULL)
}
\arguments{
\item{parameter}{}
}
\value{
what is returnts...
}
\description{
description
}
\examples{

\dontrun{
}
}
