% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_misc.R
\name{fc}
\alias{fc}
\title{Copy files and rename with date}
\usage{
fc(fpath, des = "", tag = "-")
}
\arguments{
\item{fpath}{\verb{<chr>} A vector of file paths of the source files to copy.}

\item{des}{\verb{<chr>} Destination folder.}

\item{tag}{\verb{<chr>} Tag to the filename.}
}
\value{
A logical vector indicating if the operation succeeded for each of the files.
}
\description{
Copy files to destination and add date to file names with a tag as desired.
}
\examples{
\dontrun{
# Copy a file to home directory
tmp <- tempdir()
fc(c("f1.R","f2.R"),des=tmp)
}
}
