% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{fc}
\alias{fc}
\title{Copy file to destination}
\usage{
fc(fpath, tag = "-", des = "")
}
\arguments{
\item{fpath}{\verb{<chr>} File path of the source file}

\item{tag}{\verb{<chr>} Tag to the filename}

\item{des}{\verb{<chr>} Destination folder}
}
\value{
A logical vector indicating if the operation succeeded for each of the files
}
\description{
Copy file to destination and add date to file name with a tag as desired.
}
\examples{
\dontrun{
# Copy a file to home directory
tmp <- tempdir()
fc("myfile.R",des=tmp)
}
}
