% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_all.R
\name{fc}
\alias{fc}
\title{Copy files and rename with date}
\usage{
fc(..., des = "", tag = "", td = TRUE)
}
\arguments{
\item{...}{\verb{<chr>} A vector of file paths of the source files to copy and rename.}

\item{des}{\verb{<chr>} Destination folder. "." to rename files at the current location.}

\item{tag}{\verb{<chr>} Tag to the filename.}

\item{td}{\verb{<lgl>} \code{TRUE} to add today (yymmdd) to the filename.}
}
\value{
A logical vector indicating if the operation succeeded for each of the files.
}
\description{
Copy files to destination and rename with date and a tag as desired.
}
\examples{
\dontrun{
# Copy a file to home directory
tmp = tempdir()
fc("f1.R","f2.R",des=tmp)
}
}
