% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filterOutlier.R
\name{filterOutlier}
\alias{filterOutlier}
\alias{filterOutlier.Date}
\alias{filterOutlier.default}
\alias{filterOutlier.factor}
\alias{filterOutlier.numeric}
\title{Minimize the spread/variablity in a column.}
\usage{
filterOutlier(x, ...)

\method{filterOutlier}{numeric}(x, replace_value = NA, ...)

\method{filterOutlier}{factor}(x, perc = 1, mx = 5, o_lab = "_Other", ...)

\method{filterOutlier}{Date}(x, replace_value = NA, ...)

\method{filterOutlier}{default}(x)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{
o
}
\description{
Minimize the spread/variablity in a column.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: 

\item \code{factor}: 

\item \code{Date}: 

\item \code{default}: 
}}
\examples{
stuff
}

