% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ft}
\alias{ft}
\title{flextable wrapper}
\usage{
ft(d, fnote = NULL, ttl = NULL, sig = 8, dig = 2, src = 0, omit = "")
}
\arguments{
\item{d}{A data frame}

\item{fnote}{<\code{chr}> Footnote}

\item{ttl}{<\code{chr}> Title}

\item{sig}{<\code{int}> Number of significant digits to compute}

\item{dig}{<\code{int}> Number of decimal places to display}

\item{src}{<\code{int}> Either 1 or 2 to add source label over 1 or 2 lines}

\item{omit}{<\code{chr}> Text to omit from the source label}
}
\value{
A flextable object
}
\description{
Sugar function for default flextable output.
}
\examples{
mtcars |> head() |> ft()
mtcars |> head() |> ft(src=1)
mtcars |> head() |> ft("Footnote")
mtcars |> head() |> ft("Footnote",src=1)
mtcars |> head() |> ft(sig=2,dig=1)
}
