% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ft}
\alias{ft}
\title{flextable wrapper}
\usage{
ft(d, fnote = NULL, ttl = NULL, src = NULL)
}
\arguments{
\item{d}{A data frame}

\item{fnote}{[optional] <chr> Footnote}

\item{ttl}{[optional] <chr> Title}

\item{src}{[optional] <int> Either 1 or 2 to add source label over 1 or 2 lines}
}
\value{
A flextable object
}
\description{
Sugar function for default flextable output.
}
\examples{
mtcars |> ft()
mtcars |> ft(src=1)
mtcars |> ft("This is a footnote.",src=1)
}
