% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ft}
\alias{ft}
\title{flextable wrapper}
\usage{
ft(d, fnote = NULL, ttl = NULL, src = 0, sig = 3, dig = 3)
}
\arguments{
\item{d}{A data frame}

\item{fnote}{[optional] <chr> Footnote}

\item{ttl}{[optional] <chr> Title}

\item{src}{[optional] <int> Either 1 or 2 to add source label over 1 or 2 lines}

\item{sig}{<int> Number of significant digits to compute}

\item{dig}{<int> Number of decimal places to display}
}
\value{
A flextable object
}
\description{
Sugar function for default flextable output.
}
\examples{
mtcars |> head() |> ft()
mtcars |> head() |> ft(src=1)
mtcars |> head() |> ft("Footnote",src=1)
mtcars |> head() |> ft(sig=2,dig=1)
}
