% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ft_def}
\alias{ft_def}
\title{flextable default}
\usage{
ft_def(font = "Calibri Light", fsize = 10, pad = 3)
}
\arguments{
\item{font}{\verb{<chr>} Font family - for \code{font.family}.}

\item{fsize}{\verb{<int>} Font size (in point) - for \code{font.size}.}

\item{pad}{\verb{<int>} Padding space around text - for \code{padding}.}
}
\value{
A list containing previous default values.
}
\description{
Sugar function to set flextable defaults.
The arguments are passed to \code{\link[flextable:set_flextable_defaults]{flextable::set_flextable_defaults()}}.
}
\examples{
\dontrun{
ft_def()
}
}
\seealso{
\code{\link[flextable:set_flextable_defaults]{flextable::set_flextable_defaults()}}.
}
