% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_misc.R
\name{ft_def}
\alias{ft_def}
\title{flextable defaults}
\usage{
ft_def(
  show = FALSE,
  font = "Calibri Light",
  fsize = 10,
  pad = 3,
  na = "",
  nan = "",
  ...
)
}
\arguments{
\item{show}{\verb{<lgl>} \code{TRUE} to show default values after the update}

\item{font}{\verb{<chr>} Font family - for \code{font.family}.}

\item{fsize}{\verb{<int>} Font size (in point) - for \code{font.size}.}

\item{pad}{\verb{<int>} Padding space around text - for \code{padding}.}

\item{na}{\verb{<chr>} A value to display instead of NA - for \code{na_str}}

\item{nan}{\verb{<chr>} A value to display instead of NaN - for \code{nan_str}}

\item{...}{Additional arguments to pass to \code{\link[flextable:set_flextable_defaults]{flextable::set_flextable_defaults()}}}
}
\value{
A list containing previous default values.
}
\description{
Sugar function to set flextable defaults.
The arguments are passed to \code{\link[flextable:set_flextable_defaults]{flextable::set_flextable_defaults()}}.
}
\examples{
\dontrun{
ft_def()
}
}
\seealso{
\code{\link[flextable:set_flextable_defaults]{flextable::set_flextable_defaults()}}.
}
