% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_gg.R
\name{ggbox}
\alias{ggbox}
\title{Box plot wrapper by categorical covariates}
\usage{
ggbox(d, var, cats, ...)
}
\arguments{
\item{d}{\verb{<dfr>} A data frame.}

\item{var}{\verb{<var>} A variable to plot as unquoted name.}

\item{cats}{\verb{<var>} Optional. Categorical variables to plot as a vector of unquoted names.}

\item{...}{List of arguments to pass to \link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot}.}
}
\value{
A ggplot object of a box plot.
}
\description{
Create box plots for a chosen variable by all discrete covariates in a dataset.
Orientation will follow the axis of the discrete variables.
Numeric variables will be dropped, except the chosen variable to plot.
}
\examples{
iris |> ggbox(Sepal.Length)
sleep |> ggbox(extra)
sleep |> ggbox(extra,group)
d <- mtcars |> dplyr::mutate(cyl=factor(cyl),gear=factor(gear),vs=factor(vs))
d |> ggbox(mpg)
d |> ggbox(mpg,c(cyl,vs))
}
