% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ggcov_box}
\alias{ggcov_box}
\title{Box plot wrapper for categorical variables}
\usage{
ggcov_box(d, x, y, ...)
}
\arguments{
\item{d}{\verb{<dfr>} A data frame}

\item{x, y}{\verb{<var>} Unquoted variable names (discrete and continuous)}

\item{...}{Other arguments to pass to \link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot}}
}
\value{
A ggplot object of a box plot
}
\description{
Sugar function to generate box plot for a discrete variable in a dataset.
Orientation will follow the axis of the discrete variable.
}
\examples{
iris |> ggcov_box(Species,Sepal.Length)
mtcars |> dplyr::mutate(cyl=factor(cyl)) |> ggcov_box(cyl,mpg)
}
