% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ggcov_hist}
\alias{ggcov_hist}
\title{Histogram wrapper for continuous covariates}
\usage{
ggcov_hist(d, cols = NULL, bins = 30, ...)
}
\arguments{
\item{d}{\verb{<dfr>} A data frame.}

\item{cols}{\verb{<chr>} Optional. A character vector of selected columns.}

\item{bins}{\verb{<int>} Number of bins.}

\item{...}{Other arguments to pass to \link[ggplot2:geom_histogram]{ggplot2::geom_histogram}.}
}
\value{
A ggplot object with histograms of numeric variables.
}
\description{
Sugar function to generate histograms for numeric variables in a dataset.
Non-numeric variables will be dropped.
}
\examples{
iris |> ggcov_hist()
iris |> ggcov_hist(c("Sepal.Width","Sepal.Length"))
}
