% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ggcov_violin}
\alias{ggcov_violin}
\title{Violin plot wrapper for categorical variables}
\usage{
ggcov_violin(d, x, y, ...)
}
\arguments{
\item{d}{\verb{<dfr>} A data frame}

\item{x, y}{\verb{<unquote name>} Varaiable names (one discrete and one continuous) to plot}

\item{...}{Other arguments to pass to \link[ggplot2:geom_violin]{ggplot2::geom_violin}}
}
\value{
A ggplot object with box plots of categorical variables
}
\description{
Sugar function to show box plots of a continous variable by a discrete variable in a dataset.
Orientation will follow the axis of the discrete variable.
}
\examples{
iris |> ggcov_violin(Species,Sepal.Length)
iris |> ggcov_violin(Species,Sepal.Length,)
mtcars |> dplyr::mutate(cyl=factor(cyl)) |> ggcov_violin(cyl,mpg)
}
