% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ggcov_violin}
\alias{ggcov_violin}
\title{Violin plot wrapper}
\usage{
ggcov_violin(d, cont, cats = NULL, ...)
}
\arguments{
\item{d}{\verb{<dfr>} A data frame}

\item{cont}{\verb{<var>} A continuous variable to plot as unquoted name}

\item{cats}{\verb{<chr>} Optional. Discrete variables to plot}

\item{...}{List of arguments to pass to \link[ggplot2:geom_violin]{ggplot2::geom_violin}}
}
\value{
A ggplot object with violin plots
}
\description{
Sugar function to generate violin plots for a continuous variable.
Orientation will follow the axis of the discrete variables.
}
\examples{
iris |> ggcov_violin(Sepal.Length)
d <- mtcars |> dplyr::mutate(cyl=factor(cyl),gear=factor(gear),vs=factor(vs))
d |> ggcov_violin(mpg)
d |> ggcov_violin(mpg,c("cyl","vs"))
}
