% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_eda.R
\name{gge_barplot}
\alias{gge_barplot}
\title{Plot categorical Variables using barplot}
\usage{
gge_barplot(df, group = NULL)
}
\arguments{
\item{df}{a data frame}

\item{group}{a string with the name of the varible(s) to group the plots}
}
\description{
Plot categorical Variables using barplot
}
