% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_regression.R
\name{gge_fit}
\alias{gge_fit}
\title{Plot fitted values}
\usage{
gge_fit(model, data, y, x, formula = NULL, n = 200, cat.values = NULL,
  model.id = NULL, show.points = T, pch.col = "#00000044",
  pch.pos = "identity", pch.col.cat = NULL, linetype = FALSE,
  line.size = 0.8, ylim = NULL, xlim = NULL, xlab = NULL, ylab = NULL,
  facets = NULL, facets.title.position = "left", facets.ncol = NULL,
  facets.scales = "fixed", title = NULL, title.position = "left",
  subtitle = NULL, footnote = NULL, legend.show.formulas = TRUE,
  legend.show.cat = FALSE, legend.position = "top", legend.omit = F,
  legend.title = NULL, legend.direction = "horizontal",
  legend.box = "horizontal", legend.ncol.colour = NULL,
  legend.ncol.fill = NULL, legend.col.title = NULL,
  legend.fill.title = NULL, pch.col.palette = c(brewer = "BrBG"),
  fill.palette = c(viridis = "B"))
}
\arguments{
\item{model}{a model object, a list of them, or a tidy data frame with summary(ies) of the estimation(s). A model object can be one produced as output of \code{\link[stats]{lm} } or \code{\link[stats]{glm}}.  If the tidy summary is provided instead of the model object it must contain the columns \code{term}, \code{estimate} (the point estimate), and std.error (standard error of the estimates). The content of the column \code{term} must match the name of the covariate associated the estimates as produced by \code{attr(terms(formula), 'term.labels')}, and \code{formula} must also be provided as a parameter. If there are more than one model in the summary table, it must have a column identifying the models.}

\item{data}{the original data set used to fit the model}

\item{y}{a string with the name of the dependent variable}

\item{x}{a string with the name of the independent variable that will be used to plotted in the x-axis.}

\item{formula}{a expression as used in \code{lm} function. It is used when the parameter \code{model} receives a tidy summary. If more than one model is provided in the table, \code{formula} must be a named list. The names must match the values in the column in the table that identify each model. The formula must also match the model. See examples.}

\item{n}{an integer, the number of points of \code{x} generated to produce the predicted/fitted values}

\item{cat.values}{named list of string vectors. The name of each element of the list (the string vectors) must match variable names in the data. The element of the string vectors must be strings with the name of the categories to use in the plot with the fitted values. To generate the fitted values, the numeric columns will be set to their mean value, except the column specified in the parameter \code{x}. The categorical values are set to their first category or the first category in alphabetic order. One can set the categorical variables to different values or use more than one category by setting this parameter \code{cat.value} as desired. For instance, suppose there is a categorical variable in the data set named education, taking the values of \code{High} or \code{Low}, which was used in the model. If \code{cat.value=NULL}, the plot with the predicted values will be fixed at \code{education=High}. One can use \code{education="Low"} by setting \code{cat.value=list(eductation="low")}. One can generate predicted values for both levels of education by setting \code{cat.value=list(eductation=c("low", "high"))}. See more examples in the documentation below.}

\item{model.id}{a string with the name of the column that identifies the model in the table smry.}

\item{show.points}{boolean, indicating if the points from the original data set must be plotted or not}

\item{pch.col}{a string or rgb code.}

\item{pch.pos}{a string with position of the points (e.g., jitter, identity). Default is identity.}

\item{pch.col.cat}{a string vector with the name of the categorical variable in the data to use as color key for the points}

\item{linetype}{boolean, if \code{TRUE}, the plot displays different line types if there is more than one fitted line. Default is \code{FALSE}}

\item{line.size}{numeric, the size of the fitted line}

\item{ylim}{two-dimensional numeric vector with the limits of the y-axis}

\item{xlim}{two-dimensional numeric vector with the limits of the x-axis}

\item{xlab}{string with text to display in the x-axis}

\item{ylab}{string with text to display in the y-axis}

\item{facets}{a string vector with the name of the categorical variable to generate the facets. The fitted values will be produced for each facet.}

\item{facets.title.position}{a string with \code{left}, \code{right}, or \code{center}. Indicates the position of the title of the facets}

\item{facets.ncol}{an integer, the number of columns of grid when using facets.}

\item{facets.scales}{used when facets are used. See \code{\link[ggplot2]{facet_wrap}}}

\item{title}{a string, the title of the plot}

\item{title.position}{a string (or an number) with \code{center} (or .5), \code{left} (or 0), or \code{right} (or 1)}

\item{subtitle}{a string, the subtitle of the plot}

\item{footnote}{a string, the footnote of the plot}

\item{legend.show.formulas}{boolean, if \code{TRUE} and more than one model is provided, the formula associated with the model is displayed in the legend}

\item{legend.show.cat}{boolean, if \code{TRUE} and both cat.values and more than one model are provided, it shows the categories used to fit the values in each model}

\item{legend.position}{a string (\code{top}, \code{bottom}, \code{left} (Default)), \code{right}}

\item{legend.omit}{boolean, if \code{TRUE} the legend is omitted}

\item{legend.title}{a string with the title of the legend}

\item{legend.direction}{see \code{\link[ggplot2]{guide_legend}}}

\item{legend.box}{see \code{\link[ggplot2]{guide_legend}}}

\item{legend.ncol.colour}{an integer indicating the number of columns to use in the legend with color code}

\item{legend.ncol.fill}{an integer indicating the number of columns to use in the legend with fill code (fitted line estimation interval)}

\item{legend.col.title}{string, the title of the legend with color code of the points. Default: \code{Categories}}

\item{legend.fill.title}{string, the title of the legend with color code of the fitted values estimated interval}

\item{pch.col.palette}{a named string vector with one element. The name must be the color palette (brewer, viridis and the value a string with the pallete available (ex: "Blues" for brewer palette, "A" for viridis, etc)}

\item{fill.palette}{same as for pch.col.palette, but it controls the color of the interval around the fitted line. It also accepts unnamed vector with string "grey" (use grey scales) or "plain" to use a single grey color}
}
\description{
This function plot the fitted value or the predicted probability given the model estimated using \code{lm} or \code{glm} and the original data.
}
