% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{gghist}
\alias{gghist}
\title{Histogram wrapper for numeric variables}
\usage{
gghist(d, bins = 30, ...)
}
\arguments{
\item{d}{\verb{<dfr>} A data frame}

\item{bins}{\verb{<int>} Number of bins}

\item{...}{Other arguments to pass to \link[ggplot2:geom_histogram]{ggplot2::geom_histogram}}
}
\value{
A ggplot object with histograms of numeric variables
}
\description{
Sugar function to show histograms of numeric variables in a dataset.
}
\examples{
iris |> gghist()
}
