% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ggsrc}
\alias{ggsrc}
\title{Add source label to ggplot}
\usage{
ggsrc(plt, span = 2, size = 8, col = "grey55", lab = NULL, omit = "")
}
\arguments{
\item{plt}{A ggplot object}

\item{span}{<\emph{num}> Number of lines: either 1 or 2}

\item{size}{<\emph{num}> Text size}

\item{col}{<\emph{chr}> Colour of the text}

\item{lab}{[optional] <\emph{chr}> Custom label to use instead of the default}

\item{omit}{[optional] <\emph{chr}> Text to omit from the label}
}
\value{
A ggplot object with the added label
}
\description{
Generate and add a source label with file path and run time to a ggplot object.
}
\examples{
\dontrun{
# A source label can be easily added to a ggplot object.
library(ggplot2)
p = ggplot(mtcars, aes(mpg, wt)) + geom_point()
p |> ggsrc()
p |> ggsrc(lab="My label")
}
}
