% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_gg.R
\name{ggvio}
\alias{ggvio}
\title{Violin plot wrapper for categorical covariates}
\usage{
ggvio(d, var, cats, nsub = TRUE, ...)
}
\arguments{
\item{d}{\verb{<dfr>} A data frame.}

\item{var}{\verb{<var>} A variable to plot as unquoted name.}

\item{cats}{\verb{<var>} Optional. Categorical variables to plot as a vector of unquoted names.}

\item{nsub}{\verb{<lgl>} Show number of observations.}

\item{...}{Additional arguments for \link[ggplot2:geom_violin]{ggplot2::geom_violin}.}
}
\value{
A ggplot object.
}
\description{
Create violin plots for a chosen variable by all discrete covariates in a dataset.
Numeric variables will be dropped, except the chosen variable to plot.
}
\examples{
d = mtcars |> dplyr::mutate(cyl=factor(cyl),gear=factor(gear),vs=factor(vs))
d |> ggvio(mpg)
d |> ggvio(mpg,nsub=FALSE)
d |> ggvio(mpg,c(cyl,vs))
}
