% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_misc.R
\name{hcln}
\alias{hcln}
\title{Generate hex colour codes for plotting}
\usage{
hcln(n, show = FALSE)
}
\arguments{
\item{n}{\verb{<int>} Number of colours to output.}

\item{show}{\verb{<lgl>} \code{TRUE} to show the output colours.}
}
\value{
Hex colour codes that can be used for plotting.
}
\description{
Create a vector of hex colour codes for the desired number of colours.
Colours are generated by splitting hue in the range \verb{[0,360]} in \link[grDevices:hcl]{grDevices::hcl}.
}
\examples{
hcln(6,FALSE)
hcln(4,TRUE)
}
