% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{label_src}
\alias{label_src}
\title{Generate source label}
\usage{
label_src(span = 2, omit = "", tz = TRUE)
}
\arguments{
\item{span}{\verb{<int>} Number of lines: either 1 or 2}

\item{omit}{\verb{<chr>} Text to omit from the label}

\item{tz}{\verb{<lgl>} \code{FALSE} to exclude time stamp}
}
\value{
A label showing the source file path with a time stamp
}
\description{
Generate a source label with file path and run time.
In interactive sessions, the function uses \code{rstudioapi} to get the file path.
It is designed to work in a script file in RStudio when running interactively.
It will return empty if run in the console directly.
}
\examples{
label_src(1)
label_src(tz=FALSE)
}
