% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_all.R
\name{label_tz}
\alias{label_tz}
\title{Generate time stamp label}
\usage{
label_tz(omit = "")
}
\arguments{
\item{omit}{\verb{<chr>} Text to omit from the label.}
}
\value{
A label with time stamp.
}
\description{
Generate a label with a time stamp indicating the run time.
}
\examples{
label_tz()
}
