% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_eda.R
\name{plot_heathist}
\alias{plot_heathist}
\title{Heat plot with histogram}
\usage{
plot_heathist(x = "", y = "", title = "", subtitle = "", xlab = "x",
  ylab = "y", breaks = 50, xlim = c(min(x), max(x)), ylim = c(min(y),
  max(y)))
}
\arguments{
\item{x}{numeric vector with values to plot}

\item{y}{numeric vector with values to plot}

\item{title}{string with the title of the plot}

\item{subtitle}{a string with subtitles of the plot}

\item{xlab}{string to be displayed in the x-axis}

\item{ylab}{string to be displayed in the y-axis}

\item{breaks}{integer, for the marginal histograms}

\item{xlim}{two dimensional numeric vector with limits of the x-axis}

\item{ylim}{two dimensional numeric vector with limits of the y-axis}
}
\description{
Heat plot with histogram
}
