% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{print.edar_balance}
\alias{print.edar_balance}
\title{Print}
\usage{
\method{print}{edar_balance}(x, digits = 2, ...)
}
\arguments{
\item{x}{an output of the function \code{balance}}

\item{digits}{integer, number of digits to display}

\item{...}{ignored}
}
\description{
Generic method to print the output of a \code{balance}
}
