% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printSingleVarStats.R
\name{printSingleVarStats}
\alias{printSingleVarStats}
\title{Create knit a singleVarStats object}
\usage{
printSingleVarStats(single_var_stat_object)
}
\arguments{
\item{x}{an object to be printed, returned by \link{singleVarStats}}
}
\description{
Create knit a singleVarStats object
}

