% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{summ_by}
\alias{summ_by}
\title{Summarise continuous variables by group}
\usage{
summ_by(d, cols = NULL, ..., pct = c(0.25, 0.75), xname = "")
}
\arguments{
\item{d}{A data frame}

\item{cols}{[optional] <\emph{chr}> A vector of column names to select}

\item{...}{[optional] <\emph{unquoted names}> Columns to group by}

\item{pct}{<\emph{num}> A vector of two indicating the percentiles to compute}

\item{xname}{[optional] <\emph{chr}> Characters to omit in output column names}
}
\value{
A data frame of summarised variables
}
\description{
Summarise continuous variables by group.
Non-numeric variables will be dropped.
}
\examples{
mtcars |> summ_by()
mtcars |> summ_by(pct=c(0.1,0.9))
mtcars |> summ_by("mpg")
mtcars |> summ_by("mpg",vs)
mtcars |> summ_by("mpg",vs,am)
mtcars |> summ_by(c("mpg","disp"))
mtcars |> summ_by(c("mpg","disp"),vs)
mtcars |> summ_by(c("mpg","disp"),vs,xname="mpg_")
# grouping without column selection is possible but rarely useful in large dataset
mtcars |> summ_by(NULL,vs)
}
