% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_misc.R
\name{summ_cat}
\alias{summ_cat}
\title{Summarise categorical variables}
\usage{
summ_cat(d, pos)
}
\arguments{
\item{d}{A data frame.}

\item{pos}{\verb{<chr/int>} (name or position) Optional. Extract a chosen variable.}
}
\value{
A list containing summaries for each categorical variables.
}
\description{
Summarise all categorical variables.
Numeric variables will be dropped.
}
\examples{
iris |> summ_cat()
sleep |> summ_cat()
sleep |> summ_cat("group")
sleep |> summ_cat(1)
}
