% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{summ_cat}
\alias{summ_cat}
\title{Summarise categorical variables}
\usage{
summ_cat(d, pos = NULL)
}
\arguments{
\item{d}{A data frame}

\item{pos}{\verb{<chr>}/\verb{<int>} Position (name or integer) of the summary list to return}
}
\value{
A list containing summaries of the categorical variables
}
\description{
Summarise categorical variables. Numeric variables will be dropped.
}
\examples{
sleep |> summ_cat()
sleep |> summ_cat(1)
sleep |> summ_cat("group")
}
