% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edar_regression.R
\name{tidye}
\alias{tidye}
\title{Tidy linear model summary}
\usage{
tidye(model, digits = 4)
}
\arguments{
\item{model}{a \code{lm} of \code{glm} model}

\item{digits}{integer, the number of significant digitis to use}
}
\description{
This function uses \code{\link[broom]{tidy}} to create a data.frame with model results from \code{lm} or \code{\link[stats]{glm}} functions
}
